% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNpy.R
\name{read.npy}
\alias{read.npy}
\title{Read numpy binary file}
\usage{
read.npy(filename)
}
\arguments{
\item{filename}{filename of the npy file}
}
\value{
vector, matrix of arra
}
\description{
Read numpy binary file
}
