% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNpy.R
\name{write.npy}
\alias{write.npy}
\title{Write numpy binary file}
\usage{
write.npy(obj, filename, size = NA)
}
\arguments{
\item{obj}{R object (vector, matrix, or array)}

\item{filename}{filename of the npy file}

\item{size}{size of each element (NA for natural size)}
}
\description{
Write numpy binary file
}
